Moving the termination criteria
Got error message "Unhandled Exception: System.IndexOutOfRangeException: Index was outside the bounds of the array." for the added Sudoku instance which is corrected. The problem was that the termination criteria was met but not checked when needed. So I moved the check from the end of the main loop to the beginning.